portfolio_interview_api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.52.1
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
deepl==1.22.0
distro==1.9.0
fastapi==0.115.12
feedparser==6.0.11
frozenlist==1.6.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwipiepy==0.21.0
kiwipiepy_model==0.21.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.62
langchain-openai==0.3.18
langchain-teddynote==0.3.45
langchain-text-splitters==0.3.8
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.43
lz4==4.4.4
marshmallow==3.26.1
mmh3==4.1.0
multidict==6.4.4
mypy_extensions==1.1.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==1.82.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.2.1
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
pinecone-text==0.10.0
pip==25.1
propcache==0.3.1
protobuf==5.29.5
protoc-gen-openapiv2==0.0.1
pydantic==2.11.5
pydantic_core==2.33.2
pydantic-settings==2.9.1
PyMuPDF==1.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==78.1.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tavily-python==0.7.3
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
types-requests==2.32.0.20250515
typing_extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
